Judges QA inaccuracy fixed
