+++ frontend/src/app/components/Mentors/data.ts
@@ -1,13 +1,29 @@
-import { BACKEND_URL } from "@/app/utils/config";
+import { BACKEND_URL ,SERVER_URL} from "@/app/utils/config";
+import UserEntity from "@/app/types/userentity";
 
-export default async function FetchData() {
+let mentors: UserEntity[] = [];
+
+
+const FetchData=async ()=> {
   try {
     const response = await fetch(`${BACKEND_URL}/api/mentors?populate=image`, {
       cache: "no-store",
     });
     const data = await response.json();
-    return data;
-  } catch {
-    return {};
+    for (let mentor of data.data) {
+      let entity: UserEntity = {
+        name: mentor.attributes.name,
+        position: mentor.attributes.position,
+        image: SERVER_URL + mentor.attributes.image.data.attributes.url,
+        linkedin_url: mentor.attributes.linkedin_url,
+      };
+      mentors.push(entity);
+    }
+    return mentors;
+  } catch (error){
+    return [];
   }
 }
+
+let mentors_list = FetchData();
+export default mentors;
